Remove redundant check that leads to vulnerability where element could sneak through validation:
For example:
myJanitor.clean("<form><object onmouseover=alert(document.domain) name=_sanitized></object></form>")

will not remove anything. This is because the name ends up as a property on the node and it will cause us to skip over it immediately.

The assumption is that the check was just there to avoid doing more work than necessary but computers are fast and sanitising the node multiple times should be fine and lead to the same result. It's possible people could write custom sanitiser functions that are not idempotent but we don't think that will happen in real life.